hacky rate macro created

+++ SPC_DSP/src/sizes.rs
@@ -1,9 +1,9 @@
 pub struct Sizes;
 impl Sizes {
-    pub const VOICE_COUNT: u8 = 8;
-    pub const REGISTER_COUNT: u8 = 128;
-    pub const ECHO_HIST_SIZE: u8 = 8;
-    pub const BRR_BUF_SIZE: u8 = 12;
-    pub const EXTRA_SIZE: u8 = 16;
+    pub const VOICE_COUNT: usize = 8;
+    pub const REGISTER_COUNT: usize = 128;
+    pub const ECHO_HIST_SIZE: usize = 8;
+    pub const BRR_BUF_SIZE: usize = 12;
+    pub const EXTRA_SIZE: usize = 16;
 }
